requirements.txt
html2text
mkdocs 
mkdocs-material
pymdown-extensions
python-markdown-math
pyyaml>=3.11
pygments
#pytest
#networkx
#plotly
#matplotlib>=1.5
#mkdocs-bootswatch
